\name{SpatialVx-internal}
\alias{hoods2dSetUpLists}
\alias{LocListSetup}
\alias{MakeClusterList}
\alias{makeEmptyWave}
\alias{makeWaveNames}
\alias{inside}
\alias{CI.fun}
\alias{sig.coverage}
\alias{is.sig}
\alias{detailer}
\alias{energizer}
\alias{variogram.matrix}
\alias{initGMM}
\alias{gmmEMstep}
\alias{gmmNegLogLik}
\alias{KernelGradFUN}
\alias{grads1}
\alias{grads2}
\title{
Internal functions for the SpatialVx package
}
\description{
Not to be called by the user.
}

\keyword{ internal}
