\name{SpatialVx-internal}
\alias{calculate_FSS_from_enlarged_summed_fields}
\alias{calculate_FSSvector_from_enlarged_summed_fields}
\alias{calculate_fractions_from_enlarged_summed_field}
\alias{calculate_n05_using_bisection_from_the_summed_fields}
\alias{calculate_summed_field}
\alias{enlarge_matrix}
\alias{datagrabber}
\alias{datagrabber.SpatialVx}
\alias{datagrabber.features}
\alias{datagrabber.matched}
\alias{deltammSqCen}
\alias{deltammOrig}
\alias{FeatureMatchAnalyzer.matched}
\alias{MergeIdentifier}
\alias{hoods2dPrep}
\alias{hoods2dSetUpLists}
\alias{KernelGradFUN}
\alias{locmeasures2dPrep}
\alias{LocListSetup}
\alias{MakeClusterList}
\alias{makeEmptyWave}
\alias{makeWaveNames}
\alias{ORSS}
\alias{QlossRigid}
\alias{QcorrRigid}
\alias{inside}
\alias{CI.fun}
\alias{sig.coverage}
\alias{is.sig}
\alias{detailer}
\alias{energizer}
\alias{variogram.matrix}
\alias{initGMM}
\alias{gmmEMstep}
\alias{gmmNegLogLik}
\alias{grads1}
\alias{grads2}
\alias{plot.matchedMap}
\alias{plot.matchedNoMap}
\alias{plot.saller}
\alias{plot.SpatialVxMap}
\alias{plot.SpatialVxNoMap}
\alias{plot.variographied}
\alias{plot.wavePurifyVxFieldsMap}
\alias{plot.wavePurifyVxFieldsNoMap}
\alias{plot.wavePurifyVxStats}
\alias{print.summary.interester}
\alias{print.variographied}
\alias{shrinkgrid}
\title{
Internal functions for the SpatialVx package
}
\description{
Not to be called by the user.
}

\keyword{ internal}
