\name{SpatialVx-internal}
\alias{datagrabber}
\alias{datagrabber.SpatialVx}
\alias{datagrabber.features}
\alias{datagrabber.matched}
\alias{FeatureMatchAnalyzer.matched}
\alias{MergeIdentifier}
\alias{hoods2dPrep}
\alias{hoods2dSetUpLists}
\alias{KernelGradFUN}
\alias{locmeasures2dPrep}
\alias{LocListSetup}
\alias{MakeClusterList}
\alias{makeEmptyWave}
\alias{makeWaveNames}
\alias{inside}
\alias{CI.fun}
\alias{sig.coverage}
\alias{is.sig}
\alias{detailer}
\alias{energizer}
\alias{variogram.matrix}
\alias{initGMM}
\alias{gmmEMstep}
\alias{gmmNegLogLik}
\alias{grads1}
\alias{grads2}
\alias{plot.saller}
\alias{print.summary.interester}
\alias{shrinkgrid}
\title{
Internal functions for the SpatialVx package
}
\description{
Not to be called by the user.
}

\keyword{ internal}
